requirements.txt
poetry-core==1.9.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.31.0
python-dotenv==1.0.0
pytest==8.3.3